+++ multiqc/modules/bcl2fastq/bcl2fastq.py
@@ -23,12 +23,14 @@ class MultiqcModule(BaseMultiqcModule):
         # Collect counts by lane and sample (+source_files)
         self.bcl2fastq_bylane = dict()
         self.bcl2fastq_bysample = dict()
+        self.bcl2fastq_bysample_lane = dict()
         self.source_files = dict()
         self.split_data_by_lane_and_sample()
 
         # Filter to strip out ignored sample names
         self.bcl2fastq_bylane = self.ignore_samples(self.bcl2fastq_bylane)
         self.bcl2fastq_bysample = self.ignore_samples(self.bcl2fastq_bysample)
+        self.bcl2fastq_bysample_lane = self.ignore_samples(self.bcl2fastq_bysample_lane)
 
         # Return with Warning if no files are found
         if len(self.bcl2fastq_bylane) == 0 and len(self.bcl2fastq_bysample) == 0:
@@ -61,8 +63,7 @@ class MultiqcModule(BaseMultiqcModule):
             anchor = 'bcl2fastq-bylane',
             description = 'Number of reads per lane (with number of perfect index reads).',
             helptext = """Perfect index reads are those that do not have a single mismatch.
-                All samples of a lane are combined. Undetermined reads are treated as a third category.
-                To avoid conflicts the run ID is prepended.""",
+                All samples of a lane are combined. Undetermined reads are treated as a third category.""",
             plot = bargraph.plot(
                 self.get_bar_data_from_counts(self.bcl2fastq_bylane),
                 cats,
@@ -75,21 +76,30 @@ class MultiqcModule(BaseMultiqcModule):
         )
 
         # Add section for counts by sample
+        # get cats for per-lane tab
+        lcats = set()
+        for s_name in self.bcl2fastq_bysample_lane:
+            lcats.update(self.bcl2fastq_bysample_lane[s_name].keys())
+        lcats = sorted(list(lcats))
         self.add_section (
             name = 'Clusters by sample',
             anchor = 'bcl2fastq-bysample',
-            description = 'Number of reads per sample (with number of perfect index reads)',
+            description = 'Number of reads per sample.',
             helptext = """Perfect index reads are those that do not have a single mismatch.
                 All samples are aggregated across lanes combinned. Undetermined reads are ignored.
-                Undetermined reads are treated as a separate sample.
-                To avoid conflicts the runId is prepended.""",
+                Undetermined reads are treated as a separate sample.""",
             plot = bargraph.plot(
-                self.get_bar_data_from_counts(self.bcl2fastq_bysample),
-                cats,
+                [
+                    self.get_bar_data_from_counts(self.bcl2fastq_bysample),
+                    self.bcl2fastq_bysample_lane
+                ],
+                [cats, lcats],
                 {
                     'id': 'bcl2fastq_sample_counts',
                     'title': 'bcl2fastq: Clusters by sample',
-                    'hide_zero_cats': False
+                    'hide_zero_cats': False,
+                    'ylab': 'Number of clusters',
+                    'data_labels': ['Index mismatches', 'Counts per lane']
                 }
             )
         )
@@ -189,6 +199,9 @@ class MultiqcModule(BaseMultiqcModule):
                             "yieldQ30": 0,
                             "qscore_sum": 0
                         }
+                    if not sample in self.bcl2fastq_bysample_lane:
+                        self.bcl2fastq_bysample_lane[sample] = dict()
+                    self.bcl2fastq_bysample_lane[sample][lane] = self.bcl2fastq_data[runId][lane]["samples"][sample]["total"]
                     self.bcl2fastq_bysample[sample]["total"] += self.bcl2fastq_data[runId][lane]["samples"][sample]["total"]
                     self.bcl2fastq_bysample[sample]["total_yield"] += self.bcl2fastq_data[runId][lane]["samples"][sample]["total_yield"]
                     self.bcl2fastq_bysample[sample]["perfectIndex"] += self.bcl2fastq_data[runId][lane]["samples"][sample]["perfectIndex"]